Move function to PusherUtils
