+++ src/libs/PusherUtils.js
@@ -1,0 +1,48 @@
+import CONFIG from '../CONFIG';
+import Log from './Log';
+import NetworkConnection from './NetworkConnection';
+import * as Pusher from './Pusher/pusher';
+
+/**
+ * Abstraction around subscribing to private user channel events. Handles all logs and errors automatically.
+ *
+ * @param {String} eventName
+ * @param {String} accountID
+ * @param {Function} onEvent
+ * @param {Boolean} isChunked
+ */
+function subscribeToPrivateUserChannelEvent(eventName, accountID, onEvent, isChunked = false) {
+    const pusherChannelName = `private-encrypted-user-accountID-${accountID}${CONFIG.PUSHER.SUFFIX}`;
+
+    /**
+     * @param {Object} pushJSON
+     */
+    function logPusherEvent(pushJSON) {
+        Log.info(`[Report] Handled ${eventName} event sent by Pusher`, false, pushJSON);
+    }
+
+    function onPusherResubscribeToPrivateUserChannel() {
+        NetworkConnection.triggerReconnectionCallbacks('Pusher re-subscribed to private user channel');
+    }
+
+    /**
+     * @param {*} pushJSON
+     */
+    function onEventPush(pushJSON) {
+        logPusherEvent(pushJSON);
+        onEvent(pushJSON);
+    }
+
+    /**
+     * @param {*} error
+     */
+    function onSubscriptionFailed(error) {
+        Log.hmmm('Failed to subscribe to Pusher channel', false, {error, pusherChannelName, eventName});
+    }
+    Pusher.subscribe(pusherChannelName, eventName, onEventPush, isChunked, onPusherResubscribeToPrivateUserChannel)
+        .catch(onSubscriptionFailed);
+}
+
+export default {
+    subscribeToPrivateUserChannelEvent,
+};
