+++ src/libs/actions/Report.js
@@ -26,6 +26,7 @@ import Timers from '../Timers';
 import * as ReportActions from './ReportActions';
 import Growl from '../Growl';
 import * as Localize from '../Localize';
+import PusherUtils from '../PusherUtils';
 
 let currentUserEmail;
 let currentUserAccountID;
@@ -737,14 +738,14 @@ function subscribeToUserEvents() {
     }
 
     // Live-update a report's actions when a 'report comment' event is received.
-    Pusher.subscribeToPrivateUserChannelEvent(
+    PusherUtils.subscribeToPrivateUserChannelEvent(
         Pusher.TYPE.REPORT_COMMENT,
         currentUserAccountID,
         pushJSON => updateReportWithNewAction(pushJSON.reportID, pushJSON.reportAction, pushJSON.notificationPreference),
     );
 
     // Live-update a report's actions when a 'chunked report comment' event is received.
-    Pusher.subscribeToPrivateUserChannelEvent(
+    PusherUtils.subscribeToPrivateUserChannelEvent(
         Pusher.TYPE.REPORT_COMMENT_CHUNK,
         currentUserAccountID,
         pushJSON => updateReportWithNewAction(pushJSON.reportID, pushJSON.reportAction, pushJSON.notificationPreference),
@@ -752,12 +753,12 @@ function subscribeToUserEvents() {
     );
 
     // Live-update a report's actions when an 'edit comment' event is received.
-    Pusher.subscribeToPrivateUserChannelEvent(Pusher.TYPE.REPORT_COMMENT_EDIT,
+    PusherUtils.subscribeToPrivateUserChannelEvent(Pusher.TYPE.REPORT_COMMENT_EDIT,
         currentUserAccountID,
         pushJSON => updateReportActionMessage(pushJSON.reportID, pushJSON.sequenceNumber, pushJSON.message));
 
     // Live-update a report's actions when an 'edit comment chunk' event is received.
-    Pusher.subscribeToPrivateUserChannelEvent(
+    PusherUtils.subscribeToPrivateUserChannelEvent(
         Pusher.TYPE.REPORT_COMMENT_EDIT_CHUNK,
         currentUserAccountID,
         pushJSON => updateReportActionMessage(pushJSON.reportID, pushJSON.sequenceNumber, pushJSON.message),
@@ -765,7 +766,7 @@ function subscribeToUserEvents() {
     );
 
     // Live-update a report's pinned state when a 'report toggle pinned' event is received.
-    Pusher.subscribeToPrivateUserChannelEvent(Pusher.TYPE.REPORT_TOGGLE_PINNED,
+    PusherUtils.subscribeToPrivateUserChannelEvent(Pusher.TYPE.REPORT_TOGGLE_PINNED,
         currentUserAccountID,
         pushJSON => updateReportPinnedState(pushJSON.reportID, pushJSON.isPinned));
 }
